refactor(abi): split abi.rs into folder modules
Moves `src/namespaces/abi.rs` (1182 lines) into `src/namespaces/abi/`:

  fn_ids.rs   (82 lines)  — FN_* constants + fn_id_label
  metrics.rs  (145 lines) — RuntimeMetrics, snapshot, enable flag
  mod.rs      (~984 lines) — ValueStore, __rts_dispatch, JIT table, tests

Split criterion: group by concern, not by line count. fn_ids.rs is purely
declarative (new ops just append here); metrics.rs owns all telemetry state
plus the AtomicBool enable flag; mod.rs keeps the core dispatch + value
store plumbing (they share too much private state to separate cheaply).

No behavior change. All existing `pub(crate)` / `pub` API is preserved
verbatim via re-exports in mod.rs.
